v0.94.2
* Fixed some security issues concerning the reading of corrupt image files.
* Fixed keyboard focus loss after dialog closing in certain situations on X11.
* Minor fixes.

+++ include/emGif/emGifFileModel.h
@@ -1,7 +1,7 @@
 //------------------------------------------------------------------------------
 // emGifFileModel.h
 //
-// Copyright (C) 2004-2008,2014,2018 Oliver Hamann.
+// Copyright (C) 2004-2008,2014,2018-2019 Oliver Hamann.
 //
 // Homepage: http://eaglemode.sourceforge.net/
 //
@@ -95,7 +95,7 @@ private:
 		int X,Y,Width,Height;
 		int MinCodeSize;
 		int ColorCount;
-		int DataSize,DataFill;
+		size_t DataSize,DataFill;
 		emColor * Colors;
 		char * Data;
 	};
